+++ src/test/convertbits_tests.cpp
@@ -4,7 +4,7 @@
 
 #include <utilstrencodings.h>
 #include <test/test_pivx.h>
-#include "sapling/util.h"
+#include "sapling/sapling_util.h"
 
 #include <boost/test/unit_test.hpp>
