+++ src/test/librust/sapling_keystore_tests.cpp
@@ -6,7 +6,7 @@
 #include "test/test_pivx.h"
 
 #include "sapling/address.hpp"
-#include "sapling/util.h"
+#include "sapling/sapling_util.h"
 #include <boost/test/unit_test.hpp>
 #include <univalue.h>
