+++ src/test/librust/zip32_tests.cpp
@@ -5,7 +5,7 @@
 
 #include "test/test_pivx.h"
 
-#include "sapling/util.h"
+#include "sapling/sapling_util.h"
 #include "sapling/zip32.h"
 #include <boost/test/unit_test.hpp>
