+++ src/wallet/test/wallet_test_fixture.cpp
@@ -6,7 +6,7 @@
 #include "wallet/test/wallet_test_fixture.h"
 
 #include "rpc/server.h"
-#include "sapling/util.h"
+#include "sapling/sapling_util.h"
 #include "wallet/db.h"
 #include "wallet/wallet.h"
 #include "wallet/rpcwallet.h"
